sports_api_project/requirements.txt
django 
djangorestframework 
premier-league
django-environ
psycopg2-binary
dj-database-url
requests